Drain pending-claim retention and split the mobile navigation tests
Route retainPendingClaim through a tracked retention slot on the claim
coordinator so a wallet reset drains the write instead of racing it, and
skip one started after quiesce. Move the mobile navigation cases into a
@Tags(['mobile']) file with the shared fakes in test/support.
